backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.12.1
pytest==7.4.3
httpx==0.25.1
scikit-learn==1.3.2
tensorflow==2.18.1
pandas==2.1.3
numpy==1.26.1
matplotlib==3.8.1
websockets==12.0
python-dotenv==1.0.0
